Rename users.db to usersOLDidk.db
probably corrupted, search level": " to see, no level should have quotes in it. remove that and set a number. delete all "reason" which u can find by searching , if it gets corrupted the same way mine did, searching for: "reason": "a   because it will patch in when were typing verified add, this happened when a user was using bad syntax to apply cambans , and other things like room greetings